Conversão float

+++ regressao_linear_regularizado.py
@@ -45,7 +45,7 @@ class RegularizedLinearRegression():
 
 			# atualiza pesos
 			self.theta_0 -= learning_rate*gTheta_0
-			self.theta_n = self.theta_n*(1-(learning_rate*Lambda/m)) - learning_rate*gTheta_n
+			self.theta_n = self.theta_n*(1-(float(learning_rate*Lambda)/m)) - learning_rate*gTheta_n
 
 			if print_results:
 				self.prints(i)
